Refactoring of (adaptive) zne and factories. New factories.py file with simple algorithms. Deleted old zne.py and old test_zne.py. Renamed adaptive_zne.py -> zne.py Changed tests consistently.
