+++ mitiq/pyquil/tests/test_adaptive_zne.py
@@ -1,7 +1,8 @@
 # test_adaptive_zne.py
 import numpy as np
 
-from mitiq.adaptive_zne import BatchedFactory, mitigate, zne, zne_factory
+from mitiq.factories import LinearFactory
+from mitiq.zne import mitigate, zne, zne_factory
 from mitiq.pyquil.pyquil_utils import random_identity_circuit, measure, run_program, scale_noise, Program, QVM
 
 
@@ -34,8 +35,8 @@ def test_adaptive_zne_pyquil():
     rand_circ = random_identity_circuit(depth=30)
     pq = measure(rand_circ, qid=0)
 
-    fac = BatchedFactory([1.0, 2.0, 3.0])
-    params, expects = mitigate(pq, fac, scale_noise, run_program)
-    xx = fac.reduce(expects)
+    fac = LinearFactory([1.0, 2.0, 3.0])
+    instack, outstack = mitigate(pq, fac, scale_noise, run_program)
+    xx = fac.reduce(instack, outstack)
     
     assert np.isclose(xx, 1.0, atol=1.e-1)
