+++ mitiq/tests/test_factories.py
@@ -1,8 +1,7 @@
 # test_algorithms.py
 import numpy as np
 from typing import Callable
-from mitiq.adaptive_zne import Factory
-from mitiq.algorithms import RichardsonExtr, LinearExtr, PolyExtr
+from mitiq.factories import Factory, RichardsonFactory, LinearFactory, PolyFactory
 
 
 
@@ -40,21 +39,22 @@ def apply_algorithm(algorithm_class: Factory, f: Callable[[float], float], order
 def test_richardson_extr():
     """Tests the Richardson's extrapolator."""
     for f in [f_lin, f_non_lin]:
-        f_of_zero = apply_algorithm(RichardsonExtr, f)
+        f_of_zero = apply_algorithm(RichardsonFactory, f)
         assert np.isclose(f_of_zero, f(0), atol=1.e-7)
 
 def test_linear_extr():
     """Tests the linear extrapolator."""
-    f_of_zero = apply_algorithm(LinearExtr, f_lin)
+    f_of_zero = apply_algorithm(LinearFactory, f_lin)
     assert np.isclose(f_of_zero, f_lin(0), atol=1.e-7)
 
 def test_poly_extr():
     """Tests the polinomial extrapolator."""
-    f_of_zero = apply_algorithm(PolyExtr, f_lin, 1)
+    # test linear extrapolation (order=1)
+    f_of_zero = apply_algorithm(PolyFactory, f_lin, 1)
     assert np.isclose(f_of_zero, f_lin(0), atol=1.e-7)
-    # test that order=1 is not good for non-linear function
-    f_of_zero = apply_algorithm(PolyExtr, f_non_lin, 1)
+    # test that, for some non-linear functions,
+    # order=1 is bad while ored=2 is better.
+    f_of_zero = apply_algorithm(PolyFactory, f_non_lin, 1)
     assert not np.isclose(f_of_zero, f_non_lin(0), atol=1)
-    # test that order=2 is better
-    f_of_zero = apply_algorithm(PolyExtr, f_non_lin, 2)
+    f_of_zero = apply_algorithm(PolyFactory, f_non_lin, 2)
     assert np.isclose(f_of_zero, f_non_lin(0), atol=1.e-7)
